Rename file storing all public options (consitent with TS guides)
